requirements.txt
django
djangorestframework
drf-yasg
django-filter
django-ses
django-cors-headers
